requirements.txt
# requirements.txt
# Pin all versions to guarantee reproducibility.
# Install with: pip install -r requirements.txt

# ── Core ML ───────────────────────────────────────────────────────────────────
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
scikit-learn==1.4.2

# ── Data / Visualisation ──────────────────────────────────────────────────────
Pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
pandas==2.2.2

# ── App / API ─────────────────────────────────────────────────────────────────
streamlit==1.33.0
fastapi==0.111.0
uvicorn==0.29.0

# ── Config & Utilities ────────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-cov==5.0.0

# ── Notebook ──────────────────────────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.4
